Change ExprId resolution

+++ commons/zenoh-protocol/src/core/wire_expr.rs
@@ -21,7 +21,7 @@ use core::{convert::TryInto, fmt};
 use zenoh_result::{bail, ZResult};
 
 /// A numerical Id mapped to a key expression.
-pub type ExprId = u64;
+pub type ExprId = u16;
 pub const EMPTY_EXPR_ID: ExprId = 0;
 
 /// A zenoh **resource** is represented by a pair composed by a **key** and a
